+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][latest]
+### Changed
+- Remove deprecated `disabled` config option (#671)
+
 ### Added
 - Nodejs: Include source snippet in backtraces when available (#624)